refactor(GameManager): extract UnitManagementService
- Extract 171 lines of unit management logic from GameManager
- Create UnitManagementService extending BaseGameService
- Move unit operations: createUnit, moveUnit, attackUnit, fortifyUnit
- Move unit queries: getPlayerUnits, getVisibleUnits
- Update GameManager to delegate to UnitManagementService
- Reduce GameManager from 1,822 to 1,699 lines (123 line reduction)
- All tests passing with TypeScript compilation successful
